backend/requirements.txt
fastapi==0.111.0
python-dotenv
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg[binary]
pydantic<2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
google-auth==2.29.0
requests==2.32.3
pandas
numpy
scikit-learn
xgboost
joblib
web3==6.11.4
eth-account==0.10.0
setuptools<70
websockets

